projects/Agent-8_kl/requirements.txt
langchain-openai
langgraph
tavily-python
python-dotenv
requests
tweepy
pillow
openai
langsmith
google-genai
exa_py
crawl4ai
streamlit
